freshcart_ml/requirements.txt
Flask
requests
beautifulsoup4
pillow
pytesseract
pdf2image
gunicorn

